Extract editor structure panel UI
