+++ assets/js/composer-seo-staging.js
@@ -1,4 +1,4 @@
-import { generateSitemapData, resolveSiteBaseUrl } from './seo.js?v=press-system-v3.4.25';
+import { generateSitemapData, resolveSiteBaseUrl } from './seo.js?v=press-system-v3.4.26';
 
 function escapeSeoXml(str) {
   return String(str || '').replace(/[<>&'\"]/g, (char) => {
@@ -252,11 +252,11 @@ function buildDefaultIndexHtml(metaBlock, lang) {
   html += '  <meta name="viewport" content="width=device-width, initial-scale=1.0">\n\n';
   html += metaSection;
   html += '  <!-- Note: Structured data is dynamically generated by the SEO system -->\n\n';
-  html += '  <script src="assets/js/theme-boot.js?v=press-system-v3.4.25"></script>\n';
+  html += '  <script src="assets/js/theme-boot.js?v=press-system-v3.4.26"></script>\n';
   html += '  <link rel="stylesheet" id="theme-pack">\n';
   html += '</head>\n\n';
   html += '<body>\n';
-  html += '  <script type="module" src="assets/main.js?v=press-system-v3.4.25"></script>\n';
+  html += '  <script type="module" src="assets/main.js?v=press-system-v3.4.26"></script>\n';
   html += '</body>\n\n';
   html += '</html>\n';
   return html;
