+++ assets/js/system-updates.js
@@ -1,7 +1,7 @@
-import { mdParse } from './markdown.js?v=press-system-v3.4.25';
-import { renderPressMath } from './math-render.js?v=press-system-v3.4.25';
-import { setSafeHtml } from './safe-html.js?v=press-system-v3.4.25';
-import { t } from './i18n.js?v=press-system-v3.4.25';
+import { mdParse } from './markdown.js?v=press-system-v3.4.26';
+import { renderPressMath } from './math-render.js?v=press-system-v3.4.26';
+import { setSafeHtml } from './safe-html.js?v=press-system-v3.4.26';
+import { t } from './i18n.js?v=press-system-v3.4.26';
 import {
   compareSemver,
   isUpgradeAllowed,
@@ -10,7 +10,7 @@ import {
   normalizeSemver,
   normalizeUpgradeFrom,
   semverToTag
-} from './press-version.js?v=press-system-v3.4.25';
+} from './press-version.js?v=press-system-v3.4.26';
 import { unzipSync, strFromU8 } from './vendor/fflate.browser.js';
 
 const TEXT_EXTENSIONS = new Set([
