+++ assets/press-system.json
@@ -1,13 +1,13 @@
 {
   "schemaVersion": 1,
   "type": "press-system",
-  "version": "3.4.25",
-  "tag": "v3.4.25",
+  "version": "3.4.26",
+  "tag": "v3.4.26",
   "upgradeFrom": {
     "ranges": [
-      ">=3.4.24 <3.4.25"
+      ">=3.4.25 <3.4.26"
     ],
     "allowUnknownSource": false,
-    "message": "Update to v3.4.24 first."
+    "message": "Update to v3.4.25 first."
   }
 }
